refactor: 사용하지 않는 dto 삭제 및 폴더 위치 변경
