+++ src/main/java/com/moeasy/moeasy/service/question/MakeQuestionService.java
@@ -6,9 +6,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.moeasy.moeasy.config.response.custom.CustomErrorException;
 import com.moeasy.moeasy.dto.onboarding.OnboardingQuestionAnswerDto;
 import com.moeasy.moeasy.dto.onboarding.request.OnboardingMakeQuestionRequestDto;
+import com.moeasy.moeasy.dto.onboarding.response.QuestionResponseDto;
 import com.moeasy.moeasy.dto.quesiton.MultipleChoiceIncludeIdQuestionDto;
 import com.moeasy.moeasy.dto.quesiton.MultipleChoiceQuestionDto;
-import com.moeasy.moeasy.dto.quesiton.QuestionResponseDto;
 import com.moeasy.moeasy.dto.quesiton.ShortAnswerIncludeIdQuestionDto;
 import com.moeasy.moeasy.dto.quesiton.ShortAnswerQuestionDto;
 import com.moeasy.moeasy.service.llm.NaverCloudStudioService;
