+++ src/main/java/com/moeasy/moeasy/service/question/QrCodeService.java
@@ -7,8 +7,8 @@ import com.google.zxing.client.j2se.MatrixToImageWriter;
 import com.google.zxing.common.BitMatrix;
 import com.moeasy.moeasy.config.response.custom.CustomErrorException;
 import com.moeasy.moeasy.domain.question.Question;
-import com.moeasy.moeasy.dto.onboarding.response.QrCodeResponseDto;
 import com.moeasy.moeasy.dto.quesiton.VerifyQrCodeDto;
+import com.moeasy.moeasy.dto.quesiton.response.QrCodeResponseDto;
 import com.moeasy.moeasy.repository.question.QuestionRepository;
 import com.moeasy.moeasy.service.aws.AwsService;
 import jakarta.persistence.EntityNotFoundException;
