+++ src/main/java/com/moeasy/moeasy/service/question/QuestionService.java
@@ -6,7 +6,6 @@ import com.moeasy.moeasy.config.response.custom.CustomErrorException;
 import com.moeasy.moeasy.domain.account.Member;
 import com.moeasy.moeasy.domain.question.Question;
 import com.moeasy.moeasy.domain.survey.Survey;
-import com.moeasy.moeasy.dto.onboarding.response.QrCodeResponseDto;
 import com.moeasy.moeasy.dto.quesiton.MultipleChoiceIncludeIdQuestionDto;
 import com.moeasy.moeasy.dto.quesiton.PatchQuestionTitleResponseDto;
 import com.moeasy.moeasy.dto.quesiton.QuestionsDto;
@@ -14,6 +13,7 @@ import com.moeasy.moeasy.dto.quesiton.ShortAnswerIncludeIdQuestionDto;
 import com.moeasy.moeasy.dto.quesiton.VerifyQrCodeDto;
 import com.moeasy.moeasy.dto.quesiton.request.PatchQuestionTitleRequestDto;
 import com.moeasy.moeasy.dto.quesiton.request.QuestionsRequestDto;
+import com.moeasy.moeasy.dto.quesiton.response.QrCodeResponseDto;
 import com.moeasy.moeasy.dto.quesiton.response.QuestionListResponseDto;
 import com.moeasy.moeasy.dto.quesiton.response.QuestionsResponseDto;
 import com.moeasy.moeasy.dto.survey.QuestionAnswerDto;
